Fault-tolerant database creation
